Override MediaWiki session name with standard PHP name
SimpleSamlPhp uses PHP `session.name` setting for it’s internal
sessions, even when you try to override it with the SSP
`session.cookie.name` setting. In order to let SSP find back the
session it made earlier, MediaWiki should load the session with the
same name that SSP used earlier.

This solution is not perfect; there can be conflicts between SSP and
MediaWiki variables, but it will have to do. Unless there is a better
way to handle this in SimpleSamlPhp.