Add rudimentary support to install SimpleSamlAuth with composer.
Adding name, type, and autoloader to composer.json, as well as using
globals explicitely when registering the extension were enough to
make it work. It might be a good idea to add description, licence,
etc to composer.json as well.

Basic documentation on the topic:
* https://www.mediawiki.org/wiki/Composer/For_extensions
* http://www.bn2vs.com/blog/2013/11/24/introduction-to-composer-for-mediawiki-developers/
* https://getcomposer.org/doc/02-libraries.md

The final goal of this change is to publish SimpleSamlAuth on
packagist.org so that it found by composer automatically.