Fix a bug where the session name would not be set correctly.
If SimpleSAMLphp and MediaWiki both use PHP sessions,
the standard settings will cause a conflict in the session name.
To resolve this, MediaWiki must sync its session name to
the same name that SimpleSAMLphp uses.

It seems that only `ini_get` is not enough to find the correct name.
In order to fix the value we get from `ini_get`,
we call `ini_restore` first.