Fix issue with multiple acceptable SAML attribute values
When multiple acceptable attribute values were specified, the final value
was the only that mattered. If the user had a match, then they were set to
the mediawiki group. If they were not, they were removed from the group.
This change checks to see if there are any matches and grants the group if
there are. If there are no matches then the mediawiki group is removed.